Refs #32752 - Move setting editing handling into the cell
Instead of having the table handle the modal and passing down the
callback to trigger it, move the editing logic into the inner cell.
