+++ webpack/assets/javascripts/react_app/components/SettingsTable/index.js
@@ -1,29 +1,17 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import { selectSettingsByCategory } from '../SettingRecords/SettingRecordsSelectors';
 
-import { setSettingEditing } from '../SettingRecords/SettingRecordsActions';
-
 import SettingsTable from './SettingsTable';
-
-import useSettingModal from '../SettingUpdateModal/useSettingModal';
 
 const WrappedSettingsTable = props => {
   const settings = useSelector(state =>
     selectSettingsByCategory(props.category)(state)
   );
 
-  const dispatch = useDispatch();
-  const { setModalOpen } = useSettingModal();
-
-  const onEditClick = async setting => {
-    await dispatch(setSettingEditing(setting));
-    setModalOpen();
-  };
-
-  return <SettingsTable settings={settings} onEditClick={onEditClick} />;
+  return <SettingsTable settings={settings} />;
 };
 
 WrappedSettingsTable.propTypes = {
